Move `EcPoint` to `ergo_chain_types`
