+++ ergo-chain-types/Cargo.toml
@@ -24,6 +24,8 @@ proptest-derive = {version = "0.3.0", optional = true }
 serde = { version = "1.0", features = ["derive"], optional = true }
 serde_json = { version = "1.0", optional = true }
 url = "2.2"
+k256 = { version = "0.10.2", features = ["arithmetic", "ecdsa"] }
+elliptic-curve = {version = "0.11.6", features = [ "ff"]}
 
 [dependencies.serde_with]
 version = "1.9.1"
